pyproject.toml
[tool.poetry]
name = "prellm"
version = "0.3.3"
description = "preLLM — One function for small LLM preprocessing before large LLM execution. Like litellm.completion() but with decomposition."
authors = [
    "Softreck <tom@sapletta.com>",
    "Tom Sapletta <tom@sapletta.com>",
]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/wronai/prellm"
keywords = ["llm", "prompt-engineering", "prompt-decomposition", "small-llm", "devops", "litellm"]
packages = [{include = "prellm"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
]

[tool.poetry.dependencies]
python = "^3.10"
litellm = "^1.40"
pydantic = "^2.0"
pyyaml = "^6.0"
jinja2 = "^3.1"
typer = {version = "^0.12", extras = ["all"]}
fastapi = "^0.115"
uvicorn = {version = "^0.32", extras = ["standard"]}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-mock = "^3.12"
pytest-cov = "^4.0"
ruff = "^0.4"
httpx = "^0.27"

[tool.poetry.scripts]
prellm = "prellm.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
